tidying query_builder
